Move API util into its own package
